refactor: DSLStudy코드를 Test로 이동
