+++ src/test/kotlin/dslstudy/person/Skill.kt
@@ -1,4 +1,4 @@
-package dslstudy
+package dslstudy.person
 
 data class Skill(
     val softSkill: List<String>,
